Replace u64 with f64 for FIL amount (#190)
* replace u64 with f64 for amount

* update comment

* convert all amount to f64

+++ src/cli/commands/crossmsg/fund.rs
@@ -50,6 +50,6 @@ pub(crate) struct FundArgs {
     pub from: Option<String>,
     #[arg(long, short, help = "The subnet to fund")]
     pub subnet: String,
-    #[arg(help = "The amount to fund in FIL")]
-    pub amount: u64,
+    #[arg(help = "The amount to fund in FIL, in whole FIL")]
+    pub amount: f64,
 }
